anomaly-detection-service/requirements.txt
faust-streaming==0.10.14
numpy==1.26.4
scipy==1.12.0
python-dotenv==1.0.1

